Rough initial proof-of-concept commit
The idea here is to be able to delegate license negotiation to the
sender application over the IPC connection, for cases where it is
impossible or impractical to handle it in the receiver. The basic
problem is that we cannot completely override the network request
with our own logic, so instead we take advantage of data URI support
by encoding the relevant data into a JSON map embedded in a data URI,
and overriding the "process data" handler, which can return a Promise.
Then, we use the embedded data to handle the request ourselves.

We don't yet actually use the IPC channel to perform the request,
because we've been using Chromecast Device Emulator for testing,
and that only supports replaying recorded commands. The next step
will be to embed a WebSocket server in the experimental-server.ts
to be able to test IPC delegation support.